refactor: move currency resolution to RatesController
Replace the CurrencySelection enum with RatesController.resolveInitialBalance and a giveable array extension on ExchangedBalance.
