+++ Flipcash/Core/Screens/Main/GiveViewModel.swift
@@ -94,11 +94,7 @@ final class GiveViewModel {
     init(container: Container, sessionContainer: SessionContainer, mint: PublicKey?) {
         let session          = sessionContainer.session
         let ratesController  = sessionContainer.ratesController
-        let resolved         = CurrencySelection.resolveInitialBalance(
-            mint: mint,
-            session: session,
-            ratesController: ratesController
-        )
+        let resolved         = ratesController.resolveInitialBalance(mint: mint, session: session)
 
         self.container        = container
         self.sessionContainer = sessionContainer
