+++ Flipcash/Core/Screens/Send/SendAmountViewModel.swift
@@ -91,11 +91,7 @@ final class SendAmountViewModel {
     ) {
         let session          = sessionContainer.session
         let ratesController  = sessionContainer.ratesController
-        let resolved         = CurrencySelection.resolveInitialBalance(
-            mint: mint,
-            session: session,
-            ratesController: ratesController
-        )
+        let resolved         = ratesController.resolveInitialBalance(mint: mint, session: session)
 
         self.session         = session
         self.ratesController = ratesController
